Move channel category code into its own files
