+++ store/sqlstore/channel_store_categories_test.go
@@ -1,0 +1,14 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+package sqlstore
+
+import (
+	"testing"
+
+	"github.com/mattermost/mattermost-server/v5/store/storetest"
+)
+
+func TestChannelStoreCategories(t *testing.T) {
+	StoreTestWithSqlSupplier(t, storetest.TestChannelStoreCategories)
+}
